refactor: move BaseViewModel and BaseRecyclerViewAdapter into viewmodel/adapter package
