+++ app/src/main/java/com/mashup/torchlight/ext/RecyclerViewExt.kt
@@ -2,7 +2,7 @@ package com.mashup.torchlight.ext
 
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.mashup.torchlight.ui.base.BaseRecyclerViewAdapter
+import com.mashup.torchlight.adapter.BaseRecyclerViewAdapter
 
 @BindingAdapter(value = ["replaceAll"])
 fun RecyclerView.replaceAll(list: List<Any>?) {
