+++ app/src/main/java/com/mashup/torchlight/ui/main/MainViewModel.kt
@@ -1,6 +1,6 @@
 package com.mashup.torchlight.ui.main
 
-import com.mashup.torchlight.ui.base.BaseViewModel
+import com.mashup.torchlight.viewmodel.BaseViewModel
 
 class MainViewModel : BaseViewModel() {
 }
